move helper code into separate file to increase readability
